moved bcs and delted old files
