+++ tests/boundary_conditions/test_bcs_get_boundary_dofs.py
@@ -2,7 +2,7 @@
 
 import dolfinx
 from mpi4py import MPI
-from fenicsxconcrete.experimental_setup.bcs import BoundaryConditions
+from fenicsxconcrete.boundary_conditions.bcs import BoundaryConditions
 
 
 def test():
