+++ tests/boundary_conditions/test_boundary_plane_at.py
@@ -4,7 +4,7 @@ import dolfinx
 import numpy as np
 from mpi4py import MPI
 from petsc4py.PETSc import ScalarType
-from fenicsxconcrete.experimental_setup.boundary import plane_at
+from fenicsxconcrete.boundary_conditions.boundary import plane_at
 
 
 def test():
