+++ tests/finite_element_problem/test_linear_cantilever_beam.py
@@ -2,7 +2,6 @@ import os
 import pytest
 from pathlib import Path
 
-from fenicsxconcrete.helper import Parameters
 from fenicsxconcrete.unit_registry import ureg
 from fenicsxconcrete.experimental_setup.cantilever_beam import CantileverBeam
 from fenicsxconcrete.finite_element_problem.linear_elasticity import LinearElasticity
@@ -27,7 +26,7 @@ def test_linear_cantilever_beam(dimension):
         if file.is_file():
             os.remove(file)
 
-    setup_parameters = Parameters()
+    setup_parameters = {}
     setup_parameters['length'] = 1 * ureg('m')
     setup_parameters['height'] = 0.3 * ureg('m')
     setup_parameters['width'] = 0.3 * ureg('m')  # only relevant for 3D case
@@ -36,7 +35,7 @@ def test_linear_cantilever_beam(dimension):
     setup_parameters['num_elements_height'] = 3 * ureg('')
     setup_parameters['num_elements_width'] = 3 * ureg('')  # only relevant for 3D case
 
-    fem_parameters = Parameters()
+    fem_parameters = {}
     fem_parameters['rho'] = 7750 * ureg('kg/m^3')
     fem_parameters['E'] = 210e9 * ureg('N/m^2')
     fem_parameters['nu'] = 0.28 * ureg('')
